refactor(cli-plan-run): split runner into wave submodule
Move wave/submit machinery (run_wave, submit, submit_one,
collect_pending_in_wave_order, group_by_wave, transition_body) and
the unit tests into a new private `wave` module so `runner.rs`
drops well below the 300-line cap and has headroom for further fixes.

Pure restructure: no behavior change. Identical tests still pass.

Addresses audit finding `src/runner.rs:1` (        295/300refactor LOW
lines) from docs/reviews/crate-audits/convergio-cli-plan-run.md.
